requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Django==3.1.1
idna==2.10
isort==5.5.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.1
mccabe==0.6.1
numpy==1.19.2
Pillow==7.2.0
pylint==2.6.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
zbar-py==1.0.4
